+++ spring-core/src/main/java/org/springframework/aot/hint/predicate/SerializationHintsPredicates.java
@@ -14,10 +14,13 @@
  * limitations under the License.
  */
 
-package org.springframework.aot.hint;
+package org.springframework.aot.hint.predicate;
 
 import java.util.function.Predicate;
 
+import org.springframework.aot.hint.RuntimeHints;
+import org.springframework.aot.hint.SerializationHints;
+import org.springframework.aot.hint.TypeReference;
 import org.springframework.util.Assert;
 
 /**
