+++ spring-core/src/main/java/org/springframework/aot/hint/predicate/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * Predicate support for runtime hints.
+ */
+@NonNullApi
+@NonNullFields
+package org.springframework.aot.hint.predicate;
+
+import org.springframework.lang.NonNullApi;
+import org.springframework.lang.NonNullFields;
